Move phpmyadminhook over to eventing.
